+++ android/app/src/main/kotlin/com/matthiasnehlsen/lotti/FlutterHealthFitPlugin.kt
@@ -1,17 +1,17 @@
-package com.example.lotti
+package com.matthiasnehlsen.lotti
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodChannel
+import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
-import java.util.TimeZone
 
-class FlutterNativeTimezonePlugin: FlutterPlugin, MethodCallHandler {
+class FlutterHealthFitPlugin: FlutterPlugin, MethodCallHandler {
     private lateinit var channel: MethodChannel
 
     override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
-        channel = MethodChannel(binding.binaryMessenger, "flutter_native_timezone")
+        channel = MethodChannel(binding.binaryMessenger, "flutter_health_fit")
         channel.setMethodCallHandler(this)
     }
 
@@ -21,8 +21,8 @@ class FlutterNativeTimezonePlugin: FlutterPlugin, MethodCallHandler {
 
     override fun onMethodCall(call: MethodCall, result: Result) {
         when (call.method) {
-            "getLocalTimezone" -> {
-                result.success(TimeZone.getDefault().id)
+            "getPlatformVersion" -> {
+                result.success("Android ${android.os.Build.VERSION.RELEASE}")
             }
             else -> {
                 result.notImplemented()
