+++ android/app/src/main/kotlin/com/matthiasnehlsen/lotti/FlutterNativeTimezonePlugin.kt
@@ -1,17 +1,17 @@
-package com.example.lotti
+package com.matthiasnehlsen.lotti
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodChannel
-import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
+import java.util.TimeZone
 
-class FlutterHealthFitPlugin: FlutterPlugin, MethodCallHandler {
+class FlutterNativeTimezonePlugin: FlutterPlugin, MethodCallHandler {
     private lateinit var channel: MethodChannel
 
     override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
-        channel = MethodChannel(binding.binaryMessenger, "flutter_health_fit")
+        channel = MethodChannel(binding.binaryMessenger, "flutter_native_timezone")
         channel.setMethodCallHandler(this)
     }
 
@@ -21,8 +21,8 @@ class FlutterHealthFitPlugin: FlutterPlugin, MethodCallHandler {
 
     override fun onMethodCall(call: MethodCall, result: Result) {
         when (call.method) {
-            "getPlatformVersion" -> {
-                result.success("Android ${android.os.Build.VERSION.RELEASE}")
+            "getLocalTimezone" -> {
+                result.success(TimeZone.getDefault().id)
             }
             else -> {
                 result.notImplemented()
