+++ android/app/src/main/kotlin/com/matthiasnehlsen/lotti/MainActivity.kt
@@ -1,6 +1,12 @@
 package com.matthiasnehlsen.lotti
 
 import io.flutter.embedding.android.FlutterActivity
+import io.flutter.embedding.engine.FlutterEngine
 
 class MainActivity: FlutterActivity() {
-}
+    override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
+        super.configureFlutterEngine(flutterEngine)
+        flutterEngine.plugins.add(FlutterHealthFitPlugin())
+        flutterEngine.plugins.add(FlutterNativeTimezonePlugin())
+    }
+}
